+++ R/get_vulnerabilities.R
@@ -17,8 +17,6 @@
 #' @description Parse the audit data frame (obtained via \code{audit_deps}), and extract
 #' the vulnerabilities.
 #' @param audit Output from \code{audit_deps}.
-#' @importFrom purrr map_dfr map
-#' @importFrom tidyr unnest
 #' @export
 #' @examples
 #' \dontrun{
@@ -34,23 +32,24 @@
 #' }
 get_vulnerabilities = function(audit) {
   if (sum(audit$no_of_vulnerabilities, na.rm = TRUE) == 0L) {
-    return(tibble(package = character(0), version = character(0), type = character(0),
-                  oss_package = character(0), description = character(0),
-                  reference = character(0),
-                  cvss_id = character(0), cvss_title = character(0),
-                  cvss_description = character(0), cvss_score = character(0),
-                  cvss_vector = double(0), cvss_cwe = character(0),
-                  cvss_reference = character(0),
-                  no_of_vulnerabilites = integer(0)))
+    empty_tib = tibble::tibble(package = character(0), version = character(0), type = character(0),
+                               oss_package = character(0), description = character(0),
+                               reference = character(0),
+                               cvss_id = character(0), cvss_title = character(0),
+                               cvss_description = character(0), cvss_score = character(0),
+                               cvss_vector = double(0), cvss_cwe = character(0),
+                               cvss_reference = character(0),
+                               no_of_vulnerabilites = integer(0))
+    return(empty_tib)
   }
 
   audit$vulnerabilities = audit$vulnerabilities %>%
-    map(~ map_dfr(.x, ~tibble(cvss_id = .x[[1]],
-                              cvss_title = .x[[2]],
-                              cvss_description = .x[[3]],
-                              cvss_score = .x[[4]],
-                              cvss_vector = .x[[5]],
-                              cvss_cwe = .x[[6]],
-                              cvss_reference = .x[[7]])))
+    purrr::map(~ purrr::map_dfr(.x, ~tibble::tibble(cvss_id = .x[[1]],
+                                                    cvss_title = .x[[2]],
+                                                    cvss_description = .x[[3]],
+                                                    cvss_score = .x[[4]],
+                                                    cvss_vector = .x[[5]],
+                                                    cvss_cwe = .x[[6]],
+                                                    cvss_reference = .x[[7]])))
   tidyr::unnest(audit, vulnerabilities)
 }