+++ tests/testthat/test-requirements.R
@@ -1,9 +1,8 @@
 test_that("Test requirements file", {
   skip_on_cran()
 
-  # Sanity check to make sure the directory exists
-  expect_true(file.exists("requirements.txt"))
-  audit = audit_req_txt()
+  dir = system.file("extdata", "testthat", package = "oysteR", mustWork = TRUE)
+  audit = audit_req_txt(dir)
   expect_equal(ncol(audit), 8)
   expect_equal(nrow(audit), 31)
 })